requirements.txt
fastapi 
starlette==0.22.0
strawberry-graphql[debug-server]
uvicorn[standard]
deta
